pyproject.toml
[tool.poetry]
name = "movierama"
version = "0.1.0"
description = ""
authors = ["iplitharas"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "4.1.4"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
psycopg2-binary = "2.9.5"
django-extensions = "^3.2.1"
dj-rest-auth = "^2.2.5"
django-allauth = "^0.51.0"
pyyaml = "^6.0"
uritemplate = "^4.1.1"
drf-yasg = "^1.21.4"
django-crispy-forms = "^1.14.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
black = "^22.12.0"
isort = "^5.11.3"
pylint = "^2.15.9"
pre-commit = "^2.20.0"
bandit = "^1.7.4"
pylint-django = "^2.5.3"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
